BE-50, refactor: move UserBungRepository related functions to UserBungService.
